Feat: se agregaron todas las respectivas validaciones de cada tabla (sin contar las intermedias)

+++ src/models/profile.model.js
@@ -9,7 +9,7 @@ export const Profile = sequelize.define('Profile', {
     },
 
     phone: {
-        type: DataTypes.INTEGER(16),
+        type: DataTypes.INTEGER(10),
         allowNull: true
     },
